v1.1.3 - Test all users if the plugin has been enabled and someone started the plugin
This should handle RGL's edge cases where players may elect to disable the plugin outside of official matches.

Note - As long as the plugin is enabled as part of your official config, your players will be checked when the plugin starts, when the convar is enabled, or when the round starts.